Allow localized and non-localized text to exist in same payload
The body and loc-* keys are not mutually exclusive. Same with title and
title-loc-* keys. For example, you may want to send a push notification
with non-localized text for older clients that do not have the loc-*
keys in their bundle. This change allows localized and non-localized
keys to coexist in same payload.